✨ 一覧画面を作ろう２

+++ index.php
@@ -8,10 +8,11 @@ if (isset($_SESSION['id']) && $_SESSION['name']) {
     header('Location: login.php');
     exit();
 }
+$db = dbconnect();
+
 //メッセージの投稿
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $message = filter_input(INPUT_POST, 'message', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-    $db = dbconnect();
     $stmt = $db->prepare('insert into posts (message , member_id) values(?,?)');
     if (!$stmt) {
         die($db->error);
@@ -59,14 +60,29 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                     </p>
                 </div>
             </form>
-
-            <div class="msg">
-                <img src="member_picture/" width="48" height="48" alt="" />
-                <p>○○<span class="name">（○○）</span></p>
-                <p class="day"><a href="view.php?id=">2021/01/01 00:00:00</a>
-                    [<a href="delete.php?id=" style="color: #F33;">削除</a>]
-                </p>
-            </div>
+            <?php
+            $stmt = $db->prepare('select p.id , p.member_id , p.message , p.created ,
+            m.name ,m.picture from posts p, members m where m.id = p.member_id order by id desc');
+            if (!$stmt) {
+                die($db->error);
+            }
+            $success = $stmt->execute();
+            if (!$success) {
+                die($db->error);
+            }
+            $stmt->bind_result($id, $member_id, $message, $created, $name, $picture);
+            while ($stmt->fetch()) :
+            ?>
+                <div class="msg">
+                    <?php if ($picture) : ?>
+                        <img src="member_picture/<?php echo h($picture); ?>" width="48" height="48" alt="" />
+                    <?php endif; ?>
+                    <p><?php echo h($message); ?><span class="name">(<?php echo h($name); ?>)</span></p>
+                    <p class="day"><a href="view.php?id="><?php echo h($created); ?></a>
+                        [<a href="delete.php?id=" style="color: #F33;">削除</a>]
+                    </p>
+                </div>
+            <?php endwhile; ?>
         </div>
     </div>
 </body>
